publish number, tab2 submission and format, tab1 submission
Lot of changes here.
Tab 2 now has effectively all of the information that it is expected to submit (including the new function that determines publish number based on 10 previous publish numbers), and it has written and formatted this data in accordance to doc styles.

Tab 1 now searches for the footage production label that is overwritten in tab 2, finds any matches and changes their respective row colors to inert grey.