auto-find row
Added some functionality to automatically detect the lowest row with a white color background from a selection of rows at the top of tab 2. The data fills into this row now instead of whatever cell was selected -- fewer points of failure.

There is also an adjustment here to prevent closing out an episode if an ad is detected but not matched. There is also custom error text if the episode title is missing entirely.