pivot from user choice to disclosure
Instead of asking the user if they would like to use the corrective suggestion, it now runs by default. If an assumption is made, it will inform the user of this assumption to allow them to correct any incorrect matches, however unlikely. If an exact match is found (where it's not an assumption), this disclosure does not appear.